fix(#1723): move refresh_path from watch to reload
