+++ lua/nvim-tree/actions/finders/find-file.lua
@@ -3,7 +3,7 @@ local view = require "nvim-tree.view"
 local utils = require "nvim-tree.utils"
 local renderer = require "nvim-tree.renderer"
 local core = require "nvim-tree.core"
-local watch = require "nvim-tree.explorer.watch"
+local reload = require "nvim-tree.explorer.reload"
 local Iterator = require "nvim-tree.iterators.node-iterator"
 
 local M = {}
@@ -29,7 +29,7 @@ function M.fn(fname)
 
   local absolute_paths_searched = {}
 
-  watch.refresh_path(vim.fn.fnamemodify(fname, ":h"))
+  reload.refresh_path(vim.fn.fnamemodify(fname, ":h"))
 
   local found = Iterator.builder(core.get_explorer().nodes)
     :matcher(function(node)
